fix(eldr): review round 3 — bearing accumulator annotation + empty-coord test
Round 3 (CodeRabbit — trivial):

- Fix the local `windows_by_bearing` accumulator annotation `dict[int, float]` →
  `dict[float, float]`, matching the Envelope field + the float bearings from
  round 2.
- Add a test that empty `latitude=""`/`longitude=""` attrs are treated as absent
  (lat/long None, no crash) — covering the round-2 empty-string branch.

(Skipped the Ruff TRY003 nitpick on climate.py's ValueError messages: the whole
codebase uses inline exception messages, no lint gate enforces it, and
subclassing exceptions for a demo tool is over-engineering.)

64 tests green via `ws test eldr`.

+++ eldr/geometry.py
@@ -148,7 +148,7 @@ def extract_envelope(home_path: str) -> Envelope:
     wall_by_id = {w.get("id"): w for ws in walls_by_level.values() for w in ws}
 
     surfaces: list[Surface] = []
-    windows_by_bearing: dict[int, float] = {}
+    windows_by_bearing: dict[float, float] = {}
     # Track net exterior wall area per (level, category) so we can subtract openings.
     wall_area_cm2: dict[str, float] = {}      # key: wall id -> net gross area (cm^2)
     wall_category: dict[str, str] = {}         # wall id -> category
